requirements.txt
streamlit==1.24.0
pandas==2.0.3
scikit-surprise==1.1.3
fastai==2.7.12
python-dateutil==2.8.2
numpy==1.25.2
pillow==10.0.0
